317 12:18

+++ gulimall-seckill/src/main/java/com/atguigu/gulimall/seckill/vo/SeckillSkuVo.java
@@ -27,7 +27,7 @@ public class SeckillSkuVo {
     /**
      * 秒杀总量
      */
-    private BigDecimal seckillCount;
+    private Integer seckillCount;
     /**
      * 每人限购数量
      */
